+++ src/main/java/br/pucrio/les/esg_token_backend/resources/auth/controllers/AuthenticationController.java
@@ -1,6 +1,7 @@
 package br.pucrio.les.esg_token_backend.resources.auth.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -11,9 +12,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import br.pucrio.les.esg_token_backend.resources.auth.model.CreateUserDTO;
 import br.pucrio.les.esg_token_backend.resources.auth.model.LoginDTO;
 import br.pucrio.les.esg_token_backend.resources.auth.model.TokenDTO;
+import br.pucrio.les.esg_token_backend.resources.auth.services.AuthenticationService;
 import br.pucrio.les.esg_token_backend.resources.auth.services.TokenIssuerService;
+import br.pucrio.les.esg_token_backend.resources.users.model.User;
 
 @RestController
 @RequestMapping("/auth")
@@ -23,10 +27,13 @@ public class AuthenticationController {
     private AuthenticationManager authenticationManager;
 
     @Autowired
+    private AuthenticationService authenticationService;
+
+    @Autowired
     private TokenIssuerService tokenIssuerService;
 
-    @PostMapping
-    public ResponseEntity<?> auth(@RequestBody @Validated LoginDTO loginDTO) {
+    @PostMapping("/login")
+    public ResponseEntity<?> login(@RequestBody @Validated LoginDTO loginDTO) {
         UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(
                 loginDTO.getUsername(), loginDTO.getPassword());
 
@@ -37,6 +44,17 @@ public class AuthenticationController {
         token.setToken(jwt);
         token.setType("Bearer");
         return ResponseEntity.ok(token);
+
+    }
+
+    @PostMapping("/sign-up")
+    public ResponseEntity<?> signUp(@RequestBody CreateUserDTO createUserDTO) {
+        try {
+            User user = authenticationService.createUser(createUserDTO);
+            return ResponseEntity.ok(user);
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
 
     }
 }